[Web Catalog] Use precompiled vite build assets instead of invoking a NodeJS subprocess (#4)
The previous prototype involved invoking npm build as a subprocess when generating static site files in web_export.py.

This is suboptimal -- it doesn't play well with bazel builds and requires the user to install NodeJS.

To clean this up, we include the precompiled static files into client/neuralake/export/static_site/precompiled when building the package. The precompiled static site pulls from /data.json, which is expected to be copied in.